consolidate version-select into docs-page, add projectName prop
