+++ packages/docs-page/package.json
@@ -16,7 +16,7 @@
     "@hashicorp/react-head": "^3.1.2",
     "@hashicorp/react-placeholder": "^0.1.0",
     "@hashicorp/react-search": "^6.4.1",
-    "@hashicorp/react-version-select": "^0.4.0",
+    "@hashicorp/react-select-input": ">=4.x",
     "classnames": "^2.3.1",
     "fs-exists-sync": "^0.1.0",
     "gray-matter": "^4.0.3",
