--- packages/version-select/.npmignore
@@ -1,3 +1,0 @@
-docs.mdx
-props.js
-*.test.*
