--- packages/version-select/docs.mdx
@@ -1,26 +1,0 @@
----
-componentName: VersionSelect
----
-
-<p>{packageJson.description}</p>
-
-<LiveComponent>
-{`<NextRouterDebugger>
-  <VersionSelect
-    basePath={'sentinel/intro'}
-    versions={[
-      { name: 'latest', label: 'v1.11.x (latest)' },
-      { name: 'v1.10.x', label: 'v1.10.x' },
-      { name: 'v1.9.x', label: 'v1.9.x' },
-      { name: 'v1.8.x', label: 'v1.8.x' },
-      { name: 'v1.7.x', label: 'v1.7.x' },
-    ]}
-  />
-</NextRouterDebugger>`}
-</LiveComponent>
-
-<UsageDetails packageJson={packageJson} />
-
-### Props
-
-<PropsTable props={componentProps} />
