--- packages/version-select/package.json
@@ -1,23 +1,0 @@
-{
-  "name": "@hashicorp/react-version-select",
-  "description": "A component to handle selecting a version of docs to navigate to.",
-  "version": "0.4.0",
-  "author": "HashiCorp",
-  "contributors": [
-    "Kevin Wang",
-    "Bryce Kalow"
-  ],
-  "dependencies": {
-    "@hashicorp/react-select-input": ">=4.x"
-  },
-  "license": "MPL-2.0",
-  "peerDependencies": {
-    "react": ">=16.8.x"
-  },
-  "publishConfig": {
-    "access": "public"
-  },
-  "scripts": {
-    "test": "jest"
-  }
-}
